+++ Plugins/Util/Util.cpp
@@ -597,6 +597,16 @@ ArgumentStack Util::CreateDoor(ArgumentStack&& args)
     const auto facing = Services::Events::ExtractArgument<float>(args);
     const auto tag = Services::Events::ExtractArgument<std::string>(args);
 
+    int32_t appearance = -1;
+    try
+    {
+        appearance = Services::Events::ExtractArgument<int32_t>(args);
+    }
+    catch (const std::runtime_error& e)
+    {
+        LOG_WARNING("NWNX_Util_CreateDoor() called without appearance parameter, please update nwnx_util.nss");
+    }
+
     auto resRef = CResRef(strResRef);
     Vector position = {posX, posY, posZ};
 
@@ -620,6 +630,8 @@ ArgumentStack Util::CreateDoor(ArgumentStack&& args)
             pDoor->LoadDoor(&gff, &resStruct);
             pDoor->LoadVarTable(&gff, &resStruct);
             pDoor->SetPosition(position);
+            if (appearance >= 0)
+                pDoor->m_nAppearanceType = appearance;
             Utils::SetOrientation(pDoor, facing);
 
             if (!tag.empty())